+++ fuzz/corpus/directives/missing_semicolon.dart
@@ -1,0 +1,2 @@
+import 'broken.dart'
+part 'piece.dart';