+++ fuzz/corpus/graphql/malformed.dart
@@ -1,0 +1,1 @@
+final query = '''mutation Broken($id: ID! { update(id: $id) { ok }''';