+++ fuzz/corpus/graphql/operation.dart
@@ -1,0 +1,2 @@
+const query = r'''query Viewer($id: ID!) { viewer(id: $id) { id name } }''';
+client.query(query);